% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{anon.csv}
\alias{anon.csv}
\title{anon.csv opens a qualtrics csv and removes identifying data}
\usage{
anon.csv(f, cols)
}
\arguments{
\item{f}{A CSV file}

\item{cols}{Column names to additionally exclude}
}
\value{
an anonymous .CSV file
}
\description{
anon.csv opens a qualtrics csv and removes identifying data
}
\examples{
tocome
}
