% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_to_ts.R
\name{df_to_ts}
\alias{df_to_ts}
\title{Convert Data Frame to a Time Series object.}
\usage{
df_to_ts(df)
}
\arguments{
\item{df}{Data frame where the first column is a POXIXct objet or convertible.}
}
\value{
Time series indexed by the first column of df.
}
\description{
This is a random description.
}
\details{
These are random details. Will it work somehow?
I will know it soon.
}
