% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{ecor}
\alias{ecor}
\title{ecor displays a neatly formatted correlation table for a dataframe}
\usage{
ecor(data, digits = 3, method = "pearson")
}
\arguments{
\item{data}{A dataframe}

\item{digits}{Number of digits to report (default =3)}

\item{method}{correlation method, any of c("pearson", "kendall", "spearman"}
}
\value{
dataframe with correlations
}
\description{
ecor displays a neatly formatted correlation table for a dataframe
}
\examples{
ecor(df,2,"spearman")
}
