% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{fq}
\alias{fq}
\title{fq displays counts and percentages of each value in a vector}
\usage{
fq(x, dp = 1)
}
\arguments{
\item{x}{A vector}
}
\value{
dataframe with counts and percentages
}
\description{
fq displays counts and percentages of each value in a vector
}
\examples{
tocome
}
