% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{megacor}
\alias{megacor}
\title{megacor presents a correlation table with significance indicators}
\usage{
megacor(x, y = NULL, corr.method = "pearson",
  p.adjust.method = "holm", removeTriangle = c("none", "upper",
  "lower"), result = c("none", "html", "latex"))
}
\arguments{
\item{x}{A matrix or dataframe containing variables to be correlated}

\item{y}{Optionally, a second set of variables to be correlated with those in x}

\item{corr.method}{What kind of correlations should be computed? Default is "pearson", but "spearman" and "kendall" are also supported}

\item{removeTriangle}{Specifies the formatting of the output. Which section of the table to remove c("none", "upper", "lower")}

\item{result}{Format for output c("none", "html", "latex")}
}
\value{
dataframe with correlations
}
\description{
uses lsr:correlate and adapts a function from http://www.sthda.com/english/wiki/elegant-correlation-table-using-xtable-r-package
}
\examples{
megacor(df)
megacor(df[1:4],df[5:8],corr.method = "spearman", removeTriangle = "upper", result = "html")
}
