% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{nicetable}
\alias{nicetable}
\title{Nicely formatted table}
\usage{
nicetable(table)
}
\arguments{
\item{x}{A dataframe}
}
\value{
A nicely formatted dataframe
}
\description{
Displays a table with values less than 0.001 represented as "<.001". All other numeric values displayed to 3dp.
}
\examples{
tocome
}
