% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{pv}
\alias{pv}
\title{pv neatly formats the significance of a p-value}
\usage{
pv(x, correction = "none")
}
\arguments{
\item{x}{A p-value}
}
\value{
Neatly formatted p-output
}
\description{
pv neatly formats the significance of a p-value
}
\examples{
pv(dataobject$p.value)
}
