% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{pv2}
\alias{pv2}
\title{pv2 neatly formats the significance of a p-value with equals sign}
\usage{
pv2(p, method = "none", n = length(p))
}
\arguments{
\item{x}{A p-value}
}
\value{
Neatly formatted p-output
}
\description{
pv2 neatly formats the significance of a p-value with equals sign
}
\examples{
pv2(dataobject$p.value)
}
