% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{r}
\alias{r}
\title{Quick rounding
Updated 11/2/18}
\usage{
r(x, dp = 2)
}
\arguments{
\item{x}{vector to be rounded.}

\item{dp}{Number of decimal places (defaults to 2).}
}
\value{
Rounded output
}
\description{
Quick rounding
Updated 11/2/18
}
\examples{
r(4.66544645454)
r(somevector, 3)
}
