% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{round_df}
\alias{round_df}
\title{round_df function from http://jeromyanglim.tumblr.com/post/50228877196/round-numbers-in-data-frame-that-contains-non}
\usage{
round_df(x, digits = 2)
}
\arguments{
\item{x}{a dataframe.}

\item{digits}{Number of digits to round to.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
round_df function from http://jeromyanglim.tumblr.com/post/50228877196/round-numbers-in-data-frame-that-contains-non
}
\examples{
tocome
}
