% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{simpleCap}
\alias{simpleCap}
\title{SimpleCap capitalises all first letters in a string}
\usage{
simpleCap(x)
}
\arguments{
\item{x}{A string}
}
\value{
Capitalised String
}
\description{
SimpleCap capitalises all first letters in a string
}
\examples{
tocome
}
