% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{v.test}
\alias{v.test}
\title{A better t-test}
\usage{
v.test(x, y = NULL, paired = FALSE)
}
\arguments{
\item{x}{First vector to compare.}

\item{y}{Second vector to compare.}

\item{paired}{Set FALSE (default) for independent test, TRUE for paired test}
}
\value{
Neat t-test output
}
\description{
A better t-test
}
\examples{
v.test(somevar, someothervar)
}
