% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{vmode}
\alias{vmode}
\title{Shows the most common value in a vector}
\usage{
vmode(x)
}
\arguments{
\item{x}{A vector}
}
\value{
mode value within the vector
}
\description{
Shows the most common value in a vector
}
\examples{
tocome
}
