% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{vselect}
\alias{vselect}
\title{Select based on multiple matches (multiple contains)}
\usage{
vselect(data, matchvector)
}
\arguments{
\item{data}{A dataframe}

\item{matchvector}{A vector of search terms}
}
\value{
dataframe with selected cols
}
\description{
Select a subset of columns based on multiple search strings
}
\examples{
tocome
}
