% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{wtf}
\alias{wtf}
\title{wtf is stolen from the internet somewhere. Uses excel to view objects}
\usage{
wtf(x)
}
\arguments{
\item{x}{A dataframe (or similar?)}
}
\value{
opens in excel
}
\description{
wtf is stolen from the internet somewhere. Uses excel to view objects
}
\examples{
tocome
}
