% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{wtf}
\alias{wtf}
\title{WTF is in this dataframe}
\usage{
wtf(x)
}
\arguments{
\item{x}{A dataframe (or similar?)}
}
\value{
opens in excel
}
\description{
Uses excel to view R objects. Stolen from the internet somewhere.
}
\examples{
tocome
}
